+++ code/game-of-life/src/test/java/ru/unn/agile/GameOfLife/Model/MapTest.java
@@ -14,32 +14,32 @@ public class MapTest {
 
     @Test(expected = IllegalArgumentException.class)
     public void couldNotCreateMapWithIncorrectGrid() {
-        int[][] grid = {{}};
+        byte[][] grid = {{}};
         Map testMap = new Map(grid);
     }
 
     @Test(expected = IllegalArgumentException.class)
     public void couldNotCreateMapWithIncorrectGridFormat() {
-        int[][] grid = {{0, 0}, {0, 0, 0}};
+        byte[][] grid = {{0, 0}, {0, 0, 0}};
         Map testMap = new Map(grid);
     }
 
     @Test(expected = IllegalArgumentException.class)
     public void couldNotCreateMapWithIncorrectGridValues() {
-        int[][] grid = {{5, 0}, {0, 0}};
+        byte[][] grid = {{5, 0}, {0, 0}};
         Map testMap = new Map(grid);
     }
 
     @Test
     public void shouldCreateMapWithInputParameters() {
-        int[][] grid = {{1, 0, 0}, {0, 1, 0}};
+        byte[][] grid = {{1, 0, 0}, {0, 1, 0}};
         Map testMap = new Map(grid);
         assertArrayEquals(grid, testMap.getGrid());
     }
 
     @Test
     public void shouldCorrectlyCalculateSizeOfGrid() {
-        int[][] grid = {{1, 0, 0}, {0, 1, 0}};
+        byte[][] grid = {{1, 0, 0}, {0, 1, 0}};
         int[] gridSize = {2, 3};
         Map testMap = new Map(grid);
         assertArrayEquals(testMap.getSize(), gridSize);
@@ -47,49 +47,49 @@ public class MapTest {
 
     @Test(expected = IllegalArgumentException.class)
     public void couldNotReturnAliveNeighborsCountForIncorrectCoordinates() {
-        int[][] grid = {{0}};
+        byte[][] grid = {{0}};
         Map testMap = new Map(grid);
         testMap.countAliveNeighbors(1, 1);
     }
 
     @Test
     public void shouldReturnZeroForDeadGrid() {
-        int[][] grid = {{0}};
+        byte[][] grid = {{0}};
         Map testMap = new Map(grid);
         assertEquals(0, testMap.countAliveNeighbors(0, 0));
     }
 
     @Test
     public void shouldReturnCorrectValueForUpperLeftPoint() {
-        int[][] grid = {{1, 1}, {1, 1}};
+        byte[][] grid = {{1, 1}, {1, 1}};
         Map testMap = new Map(grid);
         assertEquals(3, testMap.countAliveNeighbors(0, 0));
     }
 
     @Test
     public void shouldReturnCorrectValueForUpperRightPoint() {
-        int[][] grid = {{1, 1}, {1, 1}};
+        byte[][] grid = {{1, 1}, {1, 1}};
         Map testMap = new Map(grid);
         assertEquals(3, testMap.countAliveNeighbors(1, 0));
     }
 
     @Test
     public void shouldReturnCorrectValueForLowerLeftPoint() {
-        int[][] grid = {{1, 1}, {1, 1}};
+        byte[][] grid = {{1, 1}, {1, 1}};
         Map testMap = new Map(grid);
         assertEquals(3, testMap.countAliveNeighbors(0, 1));
     }
 
     @Test
     public void shouldReturnCorrectValueForLowerRightPoint() {
-        int[][] grid = {{1, 1}, {1, 1}};
+        byte[][] grid = {{1, 1}, {1, 1}};
         Map testMap = new Map(grid);
         assertEquals(3, testMap.countAliveNeighbors(1, 1));
     }
 
     @Test
     public void shouldReturnCorrectValueForPointWith8Neighbors() {
-        int[][] grid = {{1, 1, 1}, {1, 1, 1}, {1, 1, 1}};
+        byte[][] grid = {{1, 1, 1}, {1, 1, 1}, {1, 1, 1}};
         Map testMap = new Map(grid);
         assertEquals(testMap.countAliveNeighbors(1, 1), 8);
     }
